+++ .gitignore
@@ -1,4 +1,6 @@
 test.md
 node_modules/
 ignore/
+.cache/
+.parcel-cache/
 *.bat*
