+++ index.js
@@ -1,31 +1,15 @@
+import * as luxon from 'luxon';
+import Book from './modules/Book';
+import navFunction from './modules/navLinks';
+
 const addBtn = document.getElementById('addBtn');
 const title = document.getElementById('title');
 const author = document.getElementById('author');
-const store = window.localStorage;
-
-class Book {
-  constructor(title, author) {
-    this.title = title;
-    this.author = author;
-  }
-
-    static books = [];
-
-    addNewBook() {
-      Book.books.push({ title: this.title, author: this.author });
-      store.setItem('books', JSON.stringify(Book.books));
-    }
-
-    static removeBook(title) {
-      Book.books = Book.books.filter((book) => book.title !== title);
-      store.setItem('books', JSON.stringify(Book.books));
-    }
-}
 
 function addBook(e) {
   e.preventDefault();
 
-  // update class
+  // update class state
   const titleText = `"${title.value}"`;
   const book = new Book(titleText, author.value);
   book.addNewBook();
@@ -43,58 +27,52 @@ function addBook(e) {
   const titleElem = document.createElement('p');
   titleElem.classList.add('title');
   titleElem.innerText = titleText;
+
   const authorElem = document.createElement('p');
   authorElem.classList.add('ms-2');
   authorElem.innerText = ` by ${author.value}`;
+
   const button = document.createElement('button');
   button.classList.add('ms-auto');
   button.innerText = 'Remove';
-  button.addEventListener('click', (e) => {
+
+  function bookRemoveHandler(e) {
     const liElem = e.target.parentElement;
     let title = liElem.querySelector('.title');
     title = title.innerText;
     Book.removeBook(title);
     liElem.remove();
+    // remove list border if there are no books
     if (!Book.books.length) {
       document
         .querySelector('.list')
         .classList.remove('border', 'border-dark', 'border-2');
     }
-  });
-
-  if (Book.books.length) {
-    let count = 0;
-    do {
-      document
-        .querySelector('.list')
-        .classList.add('border', 'border-dark', 'border-2');
-      count += 1;
-    } while (count === 1);
   }
+
+  button.addEventListener('click', bookRemoveHandler);
 
   li.appendChild(titleElem);
   li.appendChild(authorElem);
   li.appendChild(button);
-
   ul.appendChild(li);
+
+  // update list border
+  if (Book.books.length) {
+    document
+      .querySelector('.list')
+      .classList.add('border', 'border-dark', 'border-2');
+  }
 }
 
 addBtn.addEventListener('click', addBook);
 
-const navAnchs = document.querySelectorAll('nav ul li a');
+// enable navbar links
+const navAnchs = document.querySelectorAll('nav li a');
+navAnchs.forEach((a) => a.addEventListener('click', navFunction));
 
-navAnchs.forEach((a) => {
-  function navFunction(e) {
-    e.preventDefault();
-    const sections = document.querySelectorAll('section');
-    const linkTarget = e.target.href;
-    sections.forEach((section) => {
-      if (linkTarget.includes(section.id)) {
-        section.classList.remove('d-none');
-      } else {
-        section.classList.add('d-none');
-      }
-    });
-  }
-  a.addEventListener('click', navFunction);
-});
+// add date
+const date = document.querySelector('.date');
+// const d = new Date();
+// date.innerHTML = `${d.toUTCString()}`;
+date.innerHTML = luxon.DateTime.now();
